# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pathfinder API",
    "description": "AI-powered personalised learning path recommender",
    "version": "0.1.0"
  },
  "paths": {
    "/api/plan": {
      "post": {
        "summary": "Create Plan",
        "operationId": "create_plan_api_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plan/relax": {
      "post": {
        "summary": "Relax Plan",
        "description": "Compute ranked relaxation options for an infeasible or tight plan.",
        "operationId": "relax_plan_api_plan_relax_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRelaxRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanRelaxResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/adapt/detour": {
      "post": {
        "summary": "Insert Plan Detour",
        "description": "D6 money-shot: Splices a remediation node into the active learning path on blockage.",
        "operationId": "insert_plan_detour_api_adapt_detour_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetourRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetourResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/profile": {
      "post": {
        "summary": "Update Profile",
        "description": "Upsert display name in the profiles table on login/signup.",
        "operationId": "update_profile_api_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/profile/resume": {
      "post": {
        "summary": "Upload Resume",
        "description": "Extract verified technical skills with provenance quotes from an uploaded resume (PDF/DOCX).\nAccrues evidence in learner profile.",
        "operationId": "upload_resume_api_profile_resume_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_api_profile_resume_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileIntakeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/profile/github": {
      "post": {
        "summary": "Profile Github User",
        "description": "Analyze public repositories and tech stack for a GitHub user.\nAccrues evidence in learner profile.",
        "operationId": "profile_github_user_api_profile_github_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitHubProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileIntakeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_resume_api_profile_resume_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_resume_api_profile_resume_post"
      },
...
      "Confidence": {
        "type": "string",
        "enum": [
          "high",
          "medium",
          "low"
        ],
        "title": "Confidence",
        "description": "How much we trust a mastery estimate. Drives Stage 0 probe priority."
      },
      "DetourRequest": {
        "properties": {
          "blocked_skill_id": {
            "type": "string",
            "title": "Blocked Skill Id"
          },
          "goal": {
            "type": "string",
            "title": "Goal",
            "default": "Machine Learning Engineer"
          }
        },
        "type": "object",
        "required": [
          "blocked_skill_id"
        ],
        "title": "DetourRequest"
      },
      "DetourResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "blocked_skill_id": {
            "type": "string",
            "title": "Blocked Skill Id"
          },
          "bridge_skill_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bridge Skill Id"
          },
          "bridge_skill_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bridge Skill Name"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success",
          "blocked_skill_id"
        ],
        "title": "DetourResponse"
      },
...
      "GitHubProfileRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "GitHubProfileRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Mastery": {
        "properties": {
          "skill_id": {
            "type": "string",
            "title": "Skill Id"
          },
          "level": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Level"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence",
            "default": "low"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/MasteryEvidence"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "skill_id",
          "level"
        ],
        "title": "Mastery",
        "description": "Per-skill mastery WITH confidence -- not a single global theta.\n\nThis is what makes the \"zero-assumption\" claim real: an unverified\nresume boast carries MEDIUM confidence and gets probed or discounted,\nnever silently trusted."
      },
      "MasteryEvidence": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "quote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "source"
        ],
        "title": "MasteryEvidence",
        "description": "Provenance for a mastery claim. Stage 0 requires a quote, never a guess."
      },
      "MilestoneOut": {
        "properties": {
          "phase": {
            "$ref": "#/components/schemas/Phase"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "total_hours": {
            "type": "number",
            "title": "Total Hours"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeOut"
            },
            "type": "array",
            "title": "Nodes"
          }
        },
        "type": "object",
        "required": [
          "phase",
          "title",
          "total_hours",
          "nodes"
        ],
        "title": "MilestoneOut"
      },
      "Modality": {
        "type": "string",
        "enum": [
          "video",
          "lab",
          "reading",
          "project",
          "assessment"
        ],
        "title": "Modality"
      },
...
      "Phase": {
        "type": "string",
        "enum": [
          "foundations",
          "core",
          "advanced",
          "capstone"
        ],
        "title": "Phase"
      },
      "PlanRelaxRequest": {
        "properties": {
          "goal": {
            "type": "string",
            "maxLength": 300,
            "minLength": 2,
            "title": "Goal"
          },
          "hours_per_week": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Hours Per Week",
            "default": 10.0
          },
          "deadline_weeks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Weeks"
          },
          "budget_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Usd"
          },
          "known": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Known"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "balanced"
          }
        },
        "type": "object",
        "required": [
          "goal"
        ],
        "title": "PlanRelaxRequest"
      },
      "PlanRelaxResponse": {
        "properties": {
          "is_feasible": {
            "type": "boolean",
            "title": "Is Feasible"
          },
          "baseline_hours": {
            "type": "number",
            "title": "Baseline Hours"
          },
          "baseline_weeks": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Weeks"
          },
          "deadline_weeks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Weeks"
          },
          "hours_per_week": {
            "type": "number",
            "title": "Hours Per Week"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/RelaxationOption"
            },
            "type": "array",
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "is_feasible",
          "baseline_hours",
          "baseline_weeks",
          "deadline_weeks",
          "hours_per_week",
          "options"
        ],
        "title": "PlanRelaxResponse"
      },
      "PlanRequest": {
        "properties": {
          "goal": {
            "type": "string",
            "maxLength": 300,
            "minLength": 2,
            "title": "Goal"
          },
          "hours_per_week": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Hours Per Week",
            "default": 10.0
          },
          "deadline_weeks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Weeks"
          },
          "budget_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Usd"
          },
          "preferred_modalities": {
            "items": {
              "$ref": "#/components/schemas/Modality"
            },
            "type": "array",
            "title": "Preferred Modalities"
          },
          "known": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Known"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "balanced"
          }
        },
        "type": "object",
        "required": [
          "goal"
        ],
        "title": "PlanRequest"
      },
      "PlanResponse": {
        "properties": {
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "target_role": {
            "type": "string",
            "title": "Target Role"
          },
          "readiness_pct": {
            "type": "integer",
            "title": "Readiness Pct"
          },
          "total_hours": {
            "type": "number",
            "title": "Total Hours"
          },
          "total_cost_usd": {
            "type": "number",
            "title": "Total Cost Usd"
          },
          "weeks_required": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weeks Required"
          },
          "is_feasible": {
            "type": "boolean",
            "title": "Is Feasible"
          },
          "prerequisite_violations": {
            "type": "integer",
            "title": "Prerequisite Violations"
          },
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/MilestoneOut"
            },
            "type": "array",
            "title": "Milestones"
          },
          "gap_count": {
            "type": "integer",
            "title": "Gap Count"
          }
        },
        "type": "object",
        "required": [
          "goal",
          "target_role",
          "readiness_pct",
          "total_hours",
          "total_cost_usd",
          "weeks_required",
          "is_feasible",
          "prerequisite_violations",
          "milestones",
          "gap_count"
        ],
        "title": "PlanResponse"
      },
...
      "ProfileIntakeResponse": {
        "properties": {
          "skills": {
            "items": {
              "$ref": "#/components/schemas/Mastery"
            },
            "type": "array",
            "title": "Skills"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "skills",
          "count"
        ],
        "title": "ProfileIntakeResponse"
      },
      "ProfileResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "is_guest": {
            "type": "boolean",
            "title": "Is Guest"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "display_name",
          "is_guest"
        ],
        "title": "ProfileResponse"
      },
      "ProfileUpdateRequest": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "display_name"
        ],
        "title": "ProfileUpdateRequest"
      },
      "RelaxationOption": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "hours_saved": {
            "type": "number",
            "title": "Hours Saved"
          },
          "new_total_hours": {
            "type": "number",
            "title": "New Total Hours"
          },
          "new_weeks_required": {
            "type": "number",
            "title": "New Weeks Required"
          },
          "is_feasible": {
            "type": "boolean",
            "title": "Is Feasible"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "description",
          "hours_saved",
          "new_total_hours",
          "new_weeks_required",
          "is_feasible"
        ],
        "title": "RelaxationOption"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}